Initial support semantic highlights
Semantic source highlighting information from Idris is no longer ignored.

Atom highlights are not wrapped around the text. Instead, they live at a
different z-index, which means they can't apply text highlights. This is
why the selection doesn't touch text colors. Thus, the semantic
highlights are applied to the background only.

Because they are ugly, they are currently disabled by default in the
settings. But it's a start that can be built upon.