fix an error in the ipkg branch
the problem was that the compiler was restarted
every time between commands.
When printing the holes the package first
typechecks (a.k.a. loads the file)
and then asks the compiler for holes.
But since the compiler was restarted during the
commands it "forgot" the loaded file and couldn't
find any holes.

issue #53