Remove that error line is highlighted
I found that it makes it very hard to read the content of the line
which has the error or to see where the cursor is. This might be
dependend on the Theme used (Material UI)

The red squiggle to mark the line at the line number bar is unaffected.