Idris now restarts if stopped and a command is run.
This fixes issue #106.
In short, if the menu item `Packages->Idris->Stop Compiler` is run,
the compiler will stop, but there is currently no way of restarting it.
If a command is run after the compiler is stopped, an error message is
displayed.

Now the Idris process state is inspected before a command is sent to the
process.  If the process is not running it will be restarted before the
command is sent.