Add browse command to inspect contents of namespaces
Select the text of a namespace (e.g. Data.Vect) in the editor
and invoke `Language Idris: Browse namespace` or press `Ctrl-Alt-b`
The contents of the namespace are displayed in the message panel.

This is not ideal, but a first step for better namespace browsing support.

This PR fixes issue #86